feat: Implement no_std support for Transform API and join fan-in
- Introduced `TransformBuilder`, `StatefulTransformBuilder`, and `TransformPipeline` for single-input transforms in `no_std + alloc`.
- Moved join fan-in functionality to `aimdb-executor`, allowing runtime-specific implementations for Tokio, Embassy, and WASM.
- Created `JoinFanInRuntime`, `JoinQueue`, `JoinSender`, and `JoinReceiver` traits to abstract join queue behavior across runtimes.
- Implemented `EmbassyJoinQueue`, `TokioJoinQueue`, and `WasmJoinQueue` for respective adapters, ensuring bounded queue semantics.
- Updated `typed_api` to expose `transform_join` for multi-input joins without embedding runtime-specific types in `aimdb-core`.
- Added documentation for the new design and usage of the transform API in `no_std` environments.
