+++ aimdb-core/src/ext_macros.rs
@@ -103,19 +103,6 @@ macro_rules! impl_record_registrar_ext {
                     $crate::transform::TransformBuilder<I, T, $runtime>,
                 ) -> $crate::transform::TransformPipeline<I, T, $runtime>;
 
-            /// Multi-input reactive transform (join).
-            ///
-            /// Derives this record from multiple input records. Panics if a `.source()` or
-            /// another `.transform()` is already registered.
-            #[cfg(feature = "std")]
-            fn transform_join<F>(
-                &'a mut self,
-                build_fn: F,
-            ) -> &'a mut $crate::RecordRegistrar<'a, T, $runtime>
-            where
-                F: FnOnce(
-                    $crate::transform::JoinBuilder<T, $runtime>,
-                ) -> $crate::transform::JoinPipeline<T, $runtime>;
         }
 
         #[cfg(feature = $feature)]
@@ -188,19 +175,6 @@ macro_rules! impl_record_registrar_ext {
                 ) -> $crate::transform::TransformPipeline<I, T, $runtime>,
             {
                 self.transform_raw::<I, F>(input_key, build_fn)
-            }
-
-            #[cfg(feature = "std")]
-            fn transform_join<F>(
-                &'a mut self,
-                build_fn: F,
-            ) -> &'a mut $crate::RecordRegistrar<'a, T, $runtime>
-            where
-                F: FnOnce(
-                    $crate::transform::JoinBuilder<T, $runtime>,
-                ) -> $crate::transform::JoinPipeline<T, $runtime>,
-            {
-                self.transform_join_raw(build_fn)
             }
         }
     };
