+++ aimdb-wasm-adapter/src/lib.rs
@@ -35,6 +35,7 @@
 extern crate alloc;
 
 pub mod buffer;
+pub mod join_fanin;
 pub mod logger;
 pub mod runtime;
 pub mod time;
